+++ src/empty_poll.ml
@@ -2,7 +2,7 @@ type t = [ `Empty ]
 type Backend.t += Empty
 
 let backend = Empty
-let create () = failwith "No polling backend available"
+let create ?num_events:_ () = failwith "No polling backend available"
 let set _ _ _ = failwith "No polling backend available"
 let wait _ _ = failwith "No polling backend available"
 let clear _ = failwith "No polling backend available"